auth-service/requirements.txt
grpcio==1.65.5
grpcio-tools==1.65.5
protobuf>=4.25.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.7
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
uvloop; platform_system != 'Windows'
uvicorn>=0.30.0





